rag/tech-manual-rag/pyproject.toml
[project]
name = "tech-manual-rag"
version = "1.0.0"
description = "tech-manual-rag"
authors = [
  {name = "Jason Brazeal", email = "dev@jasonbrazeal.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie>=1.0.2",
    "elasticsearch>=8.17.2",
    "fastapi>=0.115.12",
    "google-cloud-storage>=3.1.0",
    "google-cloud-vision>=3.10.1",
    "google-genai>=1.10.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pymupdf>=1.25.4",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ipython>=9.0.2",
    "jupyterlab>=4.3.6",
    "jupytext>=1.16.7",
    "mypy>=1.15.0",
]

[project.urls]
"Author Homepage" = "https://jasonbrazeal.com"

[tool.setuptools.packages.find]
where = ["."]
include = []

[tool.ruff]
line-length = 99
indent-width = 4

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.uv]
package = false
python-downloads = "manual"
python-preference = "only-system" # don't download new python versions
compile-bytecode = true

[tool.uv.pip]
generate-hashes = true
output-file = "requirements.txt"
